Drop PolicyKit authorisation in favour of dbus rules.
PolicyKit cannot be used by a non-root process to check whether a user has priviledges
for an action.  Since we run colord as the colord user, this was resulting in all
calls to the colord API returning authorisation errors.

This should be reverted once PolicyKit can support checking from a non-root user.