libcolorhug: Handle failure to initialise GUsb in self-tests
This can fail, and does in my VMs and on the Ubuntu autopkgtest infrastructure.
Since there are other tests of libcolorhug that can usefully be run, handle this
like a failure to find a ColorHug on the USB bus, which is already non-fatal.

(cherry picked from commit b7f423facce08d99a981f13d0240404f1a298655)